Set up Jline terminal
