+++ .editor.cmd
@@ -1,1 +1,1 @@
-java -jar UMLEditor/target/UMLEditor-1.0.jar %*
+java -jar UMLEditor/target/UMLEditor-3.0.jar %*
