+++ UMLEditor/pom.xml
@@ -30,6 +30,12 @@
             <artifactId>json-simple</artifactId>
             <version>1.1.1</version>
         </dependency>
+        <!-- https://mvnrepository.com/artifact/org.jline/jline -->
+        <dependency>
+            <groupId>org.jline</groupId>
+            <artifactId>jline</artifactId>
+            <version>3.19.0</version>
+        </dependency>
     </dependencies>
 
     <build>
@@ -59,7 +65,7 @@
                     <configuration>
                         <archive>
                             <manifest>
-                                <mainClass>org.jinxs.umleditor.UMLInterface</mainClass>
+                                <mainClass>org.jinxs.umleditor.Main</mainClass>
                                 <addClasspath>true</addClasspath>
                                 <classpathPrefix>lib/</classpathPrefix>
                             </manifest>
