+++ UMLEditor/src/main/java/org/jinxs/umleditor/Main.java
@@ -1,0 +1,28 @@
+package org.jinxs.umleditor;
+
+public class Main{
+    public static void main(String[] args) {
+        
+        if (args.length == 0) {
+            UMLGUI gui = new UMLGUI();
+        }
+        else if (args.length > 0) {
+            if (args.length == 1 && args[0].equals("--cli")) {
+                
+                UMLTerminal.build();
+                
+            }
+            else {
+                System.out.print("Unrecognized argument(s): \"");
+                for (int i = 0; i < args.length; ++i) {
+                    System.out.print(args[i]);
+                    if (i + 1 < args.length) {
+                        System.out.print(" ");
+                    }
+                }
+                System.out.println("\"");
+            }
+            
+        }
+    }
+}
