+++ UMLEditor/src/main/java/org/jinxs/umleditor/saves/s1.json
@@ -1,0 +1,1 @@
+[{"relationships":[],"methods":[],"name":"class1","fields":[]}]
